pyproject.toml
[tool.poetry]
name = "captcha-gen"
version = "0.1.0"
description = ""
authors = ["x404xx <ghost.py@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
